feat: move edit member role feature to enterprise directory
