+++ apps/web/app/(app)/environments/[environmentId]/settings/members/components/EditMemberships/MembersInfo.tsx
@@ -1,5 +1,5 @@
 import MemberActions from "@/app/(app)/environments/[environmentId]/settings/members/components/EditMemberships/MemberActions";
-import MembershipRole from "@/app/(app)/environments/[environmentId]/settings/members/components/EditMemberships/MembershipRole";
+import EditMembershipRole from "@formbricks/ee/RoleSwitch/components/EditMembershipRole";
 import { isInviteExpired } from "@/app/lib/utils";
 import { TInvite } from "@formbricks/types/invites";
 import { TMember, TMembershipRole } from "@formbricks/types/memberships";
@@ -54,7 +54,7 @@ const MembersInfo = async ({
 
           <div className="ph-no-capture col-span-3 flex flex-col items-start justify-center break-all">
             {allMembers?.length > 0 && (
-              <MembershipRole
+              <EditMembershipRole
                 isAdminOrOwner={isUserAdminOrOwner}
                 memberRole={member.role}
                 memberId={!isInvitee(member) ? member.userId : ""}
