+++ apps/web/tsconfig.json
@@ -6,7 +6,8 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "../../packages/types/*.d.ts",
-    "../../packages/lib/jwt.ts"
+    "../../packages/lib/jwt.ts",
+    "../../packages/ee/RoleSwitch/components/EditMembershipRole.tsx"
   ],
   "exclude": ["../../.env"],
   "compilerOptions": {
